refactor(config): split resolver modules
